修改邀请码创建人ID为INT型

See merge request !201

+++ easy-common/easy-entity/src/main/java/com/rongyi/easy/rmmm/vo/merchant_inviteCode/InvitationCodeVO.java
@@ -20,7 +20,7 @@ public class InvitationCodeVO {
 	private int usedInviteNum;//已使用数量
 	private int invalidNum;//作废数量
 	private Date createAt;//申请时间
-	private Long createAccountId;
+	private int createAccountId;
 	private String createAccount;
 	
 	public int getId() {
@@ -59,10 +59,11 @@ public class InvitationCodeVO {
 	public void setCreateAt(Date createAt) {
 		this.createAt = createAt;
 	}
-	public Long getCreateAccountId() {
+	
+	public int getCreateAccountId() {
 		return createAccountId;
 	}
-	public void setCreateAccountId(Long createAccountId) {
+	public void setCreateAccountId(int createAccountId) {
 		this.createAccountId = createAccountId;
 	}
 	public String getCreateAccount() {
